requirements.txt
flask
numpy==1.15.0
pillow
opencv-python
